Split loader 1 (#53)
* Split of model loader and context
mostly similar to how llama.cpp did that already:
falcon_context now contains the falcon_model and a pointer to vocab
the model contains the vocab now
this allows to now conveniently load more than one context as well as loading multiple models at once

A ton of changes - if anything acts funky please report

* System prompt improvements
Structural improvements on libfalcon

* updates

* Further adapted system prints for all fine tunes

---------

+++ ggml-cuda.cu
@@ -3167,7 +3167,7 @@ void ggml_cuda_set_main_device(int main_device) {
     // we accept setting it before initialization
     g_system_gpu_status.main_device_id = main_device;
 }
-void ggml_cuda_set_vram_reserved(int vram_reserved_bytes) {
+void ggml_cuda_set_vram_reserved(int64_t vram_reserved_bytes) {
     for (int i = 0; i < GGML_CUDA_MAX_DEVICES; ++i)
     {
         g_system_gpu_status.device_vram_reserved[i] = vram_reserved_bytes;
