+++ ggml-cuda.h
@@ -21,7 +21,7 @@ typedef struct {
     size_t total_free_vram;
     size_t device_vram_free[GGML_CUDA_MAX_DEVICES];
     size_t device_vram_total[GGML_CUDA_MAX_DEVICES];
-    int device_vram_reserved[GGML_CUDA_MAX_DEVICES]; // overrides reserved vram - may be negative to force vram swapping
+    int64_t device_vram_reserved[GGML_CUDA_MAX_DEVICES]; // overrides reserved vram - may be negative to force vram swapping
     struct cudaDeviceProp device_props[GGML_CUDA_MAX_DEVICES];
 
 } GPUStatus;
@@ -34,7 +34,7 @@ bool   ggml_init_cublas(bool check_only);
 void   ggml_cuda_update_gpu_status(int device_id);
 void   ggml_cuda_print_gpu_status(const GPUStatus *status, bool print_summary);
 void   ggml_cuda_set_max_gpus(int max_gpus);
-void   ggml_cuda_set_vram_reserved(int vram_reserved);
+void   ggml_cuda_set_vram_reserved(int64_t vram_reserved);
 void   ggml_cuda_set_tensor_split_prepare(const float * tensor_split, int num_devices);
 void   ggml_cuda_set_tensor_split(const float * tensor_split);
